"scale" the series, so they all follow the same linear trend
Print the series both before and after this "scaling".
For some reason, we see 3 graph windows instead of just 2 (one is empty).

TODO: review this to make sure you're really doing what you intend to do.
TODO: do we really want to fit them all to the same trend, or do we want
to subtract the trend and do cross-correlation afterwards? Will the
cross-correlation differ for those two cases?