yfinance-test: filling empty dates works now, and the generated values seem to be close to the ones from justETF
However, the correlation values against the original ones are < 0.95,
and from the graph it seems there is a time shift.
UPDATE: not only time shift, there are more differences - where may
these be coming from? The currency conversion or some other calculations?

TODO: try to compare data in USD, without the conversion to EUR!